Add columns of video information to network dataframes

Network is supplemented with additional downloaded video information.

Usage

AddVideoData(net, youtubeAuth = NULL, ..., writeToFile = FALSE, verbose = TRUE)

add_videos(net, youtubeAuth = NULL, ..., writeToFile = FALSE, verbose = TRUE)

Arguments

net

A named list of dataframes nodes and edges generated by Create.

youtubeAuth

YouTube Authenticate object.

...

Additional parameters passed to function.

writeToFile

Logical. Write data to file. Default is FALSE.

verbose

Logical. Output additional information. Default is TRUE.

Value

Network as a named list of three dataframes containing $nodes, $edges and $videos nodes and edges include columns for additional video data.

Note

Only supports YouTube actor networks. Refer to AddVideoData.actor.youtube.