Create a credential object to access social media APIs

Authenticate creates a credential object that enables R to make authenticated calls to social media APIs. A credential object is a S3 object containing authentication related information such as an access token or key, and a class name identifying the social media that grants authentication. Authenticate is the first step of the Authenticate, Collect and Create workflow.

Refer to Authenticate.mastodon, Authenticate.youtube, Authenticate.reddit and Authenticate.web for parameters and usage.

Usage

Authenticate(socialmedia, ..., verbose = TRUE)

Arguments

socialmedia

Character string. Identifier for social media API to authenticate with. Supported social media are "mastodon", "youtube", "reddit" and "web".

...

Optional parameters to pass to functions providied by supporting R packages that are used for social media API access.

verbose

Logical. Print messages to console. Default is TRUE.