Collect data from social media for generating networks

This function collects data from social media and structures it into a dataframe that can be used for creating networks for further analysis. Collect is the second step of the Authenticate, Collect, and Create workflow.

Refer to Collect.search.mastodon, Collect.thread.mastodon, Collect.youtube, Collect.thread.reddit, Collect.listing.reddit, Collect.search.twitter, Collect.timeline.twitter and Collect.web for parameters and usage.

Collect(credential, ...)

Arguments

credential

A credential object generated from Authenticate.

...

Optional parameters to pass to functions providied by supporting R packages that are used for social media API collection.