Collect reddit thread listings from subreddits

Collects thread listings for one or more specified subreddits and structures the data into a dataframe.

# S3 method for listing.reddit
Collect(
  credential,
  endpoint,
  subreddits,
  sort = "hot",
  period = "all",
  max = 25,
  waitTime = c(3, 5),
  ua = getOption("HTTPUserAgent"),
  writeToFile = FALSE,
  verbose = FALSE,
  ...
)

collect_reddit_listings(
  subreddits,
  sort = "new",
  period = NULL,
  max = 25,
  waitTime = c(3, 5),
  ua = vsml_ua(),
  writeToFile = FALSE,
  verbose = FALSE,
  ...
)

Arguments

credential

A credential object generated from Authenticate with class name "reddit".

endpoint

API endpoint.

subreddits

Character vector. Subreddit names to collect thread listings from.

sort

Character vector. Listing thread sort order. Options are "hot", "top", "new", and "rising". Default is "hot".

period

Character vector. Listing top threads by time period. Only applicable to sort order by "top". Options are "hour", "day", "week", "month", "year" and "all". Default is "all".

max

Numeric vector. Maximum number of threads in listing to return. Default is 25.

waitTime

Numeric vector. Time range in seconds to select random wait from in-between url collection requests. Minimum is 3 seconds. Default is c(3, 5) for a wait time chosen from between 3 and 5 seconds.

ua

Character string. Override User-Agent string to use in Reddit thread requests. Default is option("HTTPUserAgent") value as set by vosonSML.

writeToFile

Logical. Write collected data to file. Default is FALSE.

verbose

Logical. Output additional information about the data collection. Default is TRUE.

...

Additional parameters passed to function. Not used in this method.

Value

A tibble object with class names "listing" and "reddit".

Note

The reddit endpoint used for collection has maximum limit of 25 per listing.

Examples

if (FALSE) {
# subreddit url to collect threads from
subreddits  <- c("datascience")

redditListing <- redditAuth |>
  Collect(endpoint = "listing", subreddits = subreddits, sort = "new", writeToFile = TRUE)
}