Create networks from social media data

This function creates networks from social media data as produced from Collect. Create is the final step of the Authenticate, Collect and Create workflow.

There are four types of networks that can be created from collected data: activity, actor, twomode or semantic.

For activity networks refer to Create.activity.mastodon, Create.activity.youtube Create.activity.reddit and Create.activity.twitter for parameters and usage.

For actor networks refer to Create.actor.mastodon, Create.actor.youtube, Create.actor.reddit and Create.actor.twitter.

For twomode and semantic networks refer to Create.twomode.twitter and Create.semantic.twitter functions for parameters and usage respectively.

Create(datasource, type, ...)

Arguments

datasource

Collected social media data of class "datasource" and socialmedia.

type

Character string. Type of network to be created, can be "activity", "actor", "twomode" or "semantic".

...

Optional parameters to pass to functions providied by supporting R packages that are used for social media network creation.