Create web activity network

Creates a web page activity network from pages. Nodes are web pages.

Usage

# S3 method for class 'activity.web'
Create(
  datasource,
  type,
  lcase = TRUE,
  ...,
  writeToFile = FALSE,
  verbose = TRUE
)

Arguments

datasource

Collected social media data with "datasource" and "web" class names.

type

Character string. Type of network to be created, set to "activity".

lcase

Logical. Convert urls and page names to lowercase.

...

Additional parameters passed to function. Not used in this method.

writeToFile

Logical. Write data to file. Default is FALSE.

verbose

Logical. Output additional information. Default is TRUE.

Value

Network as a named list of two dataframes containing $nodes and $edges.

Examples

if (FALSE) { # \dontrun{
# create a web activity network graph
net_activity <- data_collect |> Create("activity")

# network
# net_activity$nodes
# net_activity$edges
} # }