Import rtweet collected data

Imports rtweet collected data from rda or rds saved object file or from an rtweet dataframe. Ensures datasource and specified socialmedia type are set so data is usable by Create functions. Not required if collected data was collected by vosonSML and saved as an rds file, use readRDS instead.

ImportRtweet(data)

Arguments

data

Character string or dataframe. File path to or tibble of collected data from rtweet.

Value

A dataframe suitable for input into twitter network Create functions.

Note

Only supports rtweet data collected using the search_tweets or get_timeline functions.

Examples

if (FALSE) {
# import rtweet collected data from dataframe
collect_tw <- ImportRtweet(rtweet_search_data)

# import rtweet collected data from file
collect_tw <- ImportRtweet("./rtweet_search_n100.rds")
}