Merge collected data files

Merge collected data files

MergeFiles(
  path = ".",
  pattern = "(?-i).+?\\.rds$",
  unique = TRUE,
  rev = TRUE,
  writeToFile = FALSE,
  verbose = FALSE
)

merge_files(
  path = ".",
  pattern = "(?-i).+?\\.rds$",
  unique = TRUE,
  rev = TRUE,
  writeToFile = FALSE,
  verbose = FALSE
)

Arguments

path

Directory path of Collect data to merge. Default is the working directory.

pattern

Regular expression (regex) for matching file names to merge.

unique

Logical. Remove duplicates based on observation id. Default is TRUE.

rev

Logical. Reverses order of observations before removing duplicates. If collect data is provided chronologically then this should ensure the most recent copy of a duplicate is kept. Default is TRUE.

writeToFile

Logical. Save data to a file in the current working directory. Default is FALSE.

verbose

Logical. Output additional information. Default is FALSE.

Value

A merged Collect object.