Create an igraph graph from network

Create an igraph graph from network

Graph(net, directed = TRUE, writeToFile = FALSE, ...)

Arguments

net
A named list of dataframes nodes and edges generated by Create.

directed
Logical. Create a directed graph. Default is TRUE.

writeToFile
Logical. Save graph to a file in the current working directory. Default is FALSE.

...
Additional parameters passed to function. Not used in this method.

Value

An igraph object.