Collection and network analysis of social media data

The goal of the vosonSML package is to provide a suite of easy-to-use tools for collecting data from social media and generating different types of networks suited to Social Network Analysis (SNA) and text analytics. It offers tools to create unimodal, multimodal, and semantic networks. Excellent packages such as rtweet, RedditExtractoR, magrittr, dplyr and igraph were drawn on to provide an integrated work flow for creating different types of networks out of social media data. Creating networks from online social media is often non-trivial and time consuming. This package simplifies such tasks so users can focus on analysis.

vosonSML uses a straightforward S3 class system. Data collected with this package produces data.frame inheritable objects that are assigned the class "datasource". Additionally, "datasource" objects are attributed a class identifying the source of data, such as "twitter" or "youtube". In this way datasource objects are fast, easy to work with, and can be used as input to easily construct different kinds of networks. For example, the function Collect can be used to collect twitter data, which is then passed to the Create function resulting in a twitter network (as igraph object) that is ready for analysis.