vosonSML: Collecting Social Media Data and Generating Networks for Analysis

A suite of easy to use functions for collecting social media data and generating networks for analysis. Supports Twitter, Youtube, Reddit and web site data sources.

See also

Author

Maintainer: Bryan Gertzel bryan.gertzel@anu.edu.au

Authors:

Robert Ackland robert.ackland@anu.edu.au (ORCID)

Timothy Graham (ORCID)

Other contributors:

Francisca Borquez [contributor]